pyproject.toml
[tool.poetry]
name = "slackbot-release"
version = "0.1.0"
description = ""
authors = ["Jordan Lund <jlund@mozilla.com>"]

[tool.poetry.dependencies]
python = "^3.7"
slackclient = "^2.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
